+++ midap/imcut/base_cutout.py
@@ -38,6 +38,7 @@ class CutoutImage(ABC):
 
         # this should be set by the cut_corners routine
         self.corners_cut = None
+        self.offsets = None
 
         # get the file lists
         self.channels = [[os.path.join(channel, f) for f in sorted(os.listdir(channel))] for channel in self.paths]
@@ -71,16 +72,20 @@ class CutoutImage(ABC):
             shift = self.align_two_images(src, ref)
             self.shifts.append(shift)
 
-    def do_cutout(self, img, corners_cut):
+    def do_cutout(self, img, corners_cut, padding=10):
         """
         Performs a cutout of an image
         :param img: Image ad array
         :param corners_cut: The corners used for the cutout
+        :param padding: Apply this savety padding to the cutout in case the corner are outside the image
         :returns: The cutout from the image given the corners
         """
 
         # generate cutout of image
-        left_x, right_x, lower_y, upper_y = corners_cut
+        left_x, right_x, lower_y, upper_y = [c + padding for c in corners_cut]
+
+        # pad the image and cutout
+        img = np.pad(img, padding, mode='constant', constant_values=0)
         cutout = img[lower_y:upper_y, left_x:right_x]
         return cutout
 
@@ -92,19 +97,58 @@ class CutoutImage(ABC):
         """
         img_scaled = (255 * ((img - np.min(img))/np.max(img - np.min(img)))).astype('uint8')
         return img_scaled
+    
+    # def save_cutout(self, files, file_names, chamber=None, normalization=True):
+    #     """
+    #     Saves the cutouts into the proper directory
+    #     :param files: A list of arrrays (the cutouts) to save
+    #     :param file_names: The list of file names from the original files
+    #     :param chamber: The chamber number, if None, if won't be included in the path
+    #     """
+    #     # save of cutouts
+    #     # TODO: This should not be hardcoded
+    #     dir_name = os.path.dirname(os.path.dirname(file_names[0]))
+    #     for f, i in zip(file_names, files):
+    #         if normalization:
+    #             fname = f"{os.path.splitext(os.path.basename(f))[0]}_cut.png"
+    #             io.imsave(os.path.join(dir_name, 'cut_im', fname), i, check_contrast=False)
+    #         else:
+    #             fname = f"{os.path.splitext(os.path.basename(f))[0]}_cut_rawcounts.png"
+    #             io.imsave(os.path.join(dir_name, 'cut_im_rawcounts', fname), i, check_contrast=False)
 
-    def save_cutout(self, files, file_names):
+    def save_cutout(self, files, file_names, normalization, chamber=None):
         """
         Saves the cutouts into the proper directory
         :param files: A list of arrrays (the cutouts) to save
         :param file_names: The list of file names from the original files
+        :param chamber: The chamber number, if None, if won't be included in the path
         """
         # save of cutouts
         # TODO: This should not be hardcoded
         dir_name = os.path.dirname(os.path.dirname(file_names[0]))
-        for f, i in zip(file_names, files):
-            fname = f"{os.path.splitext(os.path.basename(f))[0]}_cut.png"
-            io.imsave(os.path.join(dir_name, 'cut_im', fname), i, check_contrast=False)
+        if normalization:
+            for f, i in zip(file_names, files):
+                fname = f"{os.path.splitext(os.path.basename(f))[0]}_cut.png"
+                if chamber is None:
+                    f_path = os.path.join(dir_name, 'cut_im', fname)
+                else:
+                    # we need to create this directory
+                    f_path = os.path.join(dir_name, f'chamber_{chamber}', 'cut_im')
+                    os.makedirs(f_path, exist_ok=True)
+                    f_path = os.path.join(f_path, fname)
+                io.imsave(f_path, i, check_contrast=False)
+        else:
+            for f, i in zip(file_names, files):
+                fname = f"{os.path.splitext(os.path.basename(f))[0]}_cut_rawcounts.tif"
+                if chamber is None:
+                    f_path = os.path.join(dir_name, 'cut_im_rawcounts', fname)
+                else:
+                    # we need to create this directory
+                    f_path = os.path.join(dir_name, f'chamber_{chamber}', 'cut_im_rawcounts')
+                    os.makedirs(f_path, exist_ok=True)
+                    f_path = os.path.join(f_path, fname)
+
+                io.imsave(f_path, i, check_contrast=False)
         
     def run_align_cutout(self):
         """
@@ -120,6 +164,7 @@ class CutoutImage(ABC):
             self.logger.info(f'Starting with channel {channel_id+1}/{len(self.channels)}')
             # list for the aligned cutouts
             aligned_cutouts = []
+            aligned_cutouts_norm = []
 
             # get the first image
             src = io.imread(files[0])
@@ -131,11 +176,13 @@ class CutoutImage(ABC):
 
             # perform the cutout of the first image
             cutout = self.do_cutout(src, self.corners_cut)
+            
             # scale the pixel values
             cut_src = self.scale_pixel_val(cutout)
 
             # add to list
-            aligned_cutouts.append(cut_src)
+            aligned_cutouts_norm.append(cut_src)
+            aligned_cutouts.append(cutout)
 
             # cutout of all other images of all channels
             for i in tqdm(range(1, len(files))):
@@ -150,9 +197,75 @@ class CutoutImage(ABC):
                 cut_img = self.do_cutout(img, current_corners)
                 # sacle the pixel values
                 proc_img = self.scale_pixel_val(cut_img)
-                aligned_cutouts.append(proc_img)
+                aligned_cutouts_norm.append(proc_img)
+                aligned_cutouts.append(cut_img)
 
-            self.save_cutout(aligned_cutouts, files)
+            self.save_cutout(aligned_cutouts_norm, files, normalization=True)
+            self.save_cutout(aligned_cutouts, files, normalization=False)
+
+    def run_align_cutout_mother_machine(self):
+        """
+        Aligns and cut out all images from all channels
+        """
+
+        self.logger.info('Aligning images...')
+        self.align_all_images()
+
+        self.logger.info('Cutting images...')
+        # cycle through the channels
+        for channel_id, files in enumerate(self.channels):
+            self.logger.info(f'Starting with channel {channel_id + 1}/{len(self.channels)}')
+
+            # get the first image
+            src = io.imread(files[0])
+
+            # We cut the corners if the corners_cut is None
+            if self.corners_cut is None or self.offsets is None:
+                # set the corner to cut
+                self.cut_corners(img=src)
+
+            # cut out all chambers sequentially
+            for chamber in range(0, len(self.offsets)):
+                self.logger.info(f'Starting with chamber {chamber+1}/{len(self.offsets)}')
+
+                # list for the aligned cutouts
+                aligned_cutouts = []
+                aligned_cutouts_norm = []
+
+                # adapt the corner with the shift of the image
+                base_corners = (self.corners_cut[0] + self.offsets[chamber],
+                                self.corners_cut[1] + self.offsets[chamber],
+                                self.corners_cut[2],
+                                self.corners_cut[3])
+
+                # perform the cutout of the first image
+                cutout = self.do_cutout(src, base_corners)
+                # scale the pixel values
+                cut_src = self.scale_pixel_val(cutout)
+
+                # add to list
+                aligned_cutouts_norm.append(cut_src)
+                aligned_cutouts.append(cutout)
+
+                # cutout of all other images of all channels
+                for i in tqdm(range(1, len(files))):
+                    img = io.imread(files[i])
+
+                    # adapt the corner with the shift of the image
+                    left_x, right_x, lower_y, upper_y = base_corners
+                    current_corners = (left_x - self.shifts[i - 1][1],
+                                       right_x - self.shifts[i - 1][1],
+                                       lower_y - self.shifts[i - 1][0],
+                                       upper_y - self.shifts[i - 1][0])
+
+                    cut_img = self.do_cutout(img, current_corners)
+                    # sacle the pixel values
+                    proc_img = self.scale_pixel_val(cut_img)
+                    aligned_cutouts_norm.append(proc_img)
+                    aligned_cutouts.append(cut_img)
+
+                self.save_cutout(aligned_cutouts_norm, files, normalization=True, chamber=chamber)
+                self.save_cutout(aligned_cutouts, files, normalization=False, chamber=chamber)
 
     @abstractmethod
     def cut_corners(self, img):